Edabit challenge from the Javascript hard section.
Build a function that creates histograms. 
Every bar needs to be on a new line and its length corresponds to the numbers in the array passed as an argument. 
The second argument of the function represents the character to be used for the bar.

histogram(arr, char) ➞ str
Examples
histogram([1, 3, 4], "#") ➞ "#\n###\n####"